decimal to float

+++ backend/db/migrations/20240924034225-create-spot.js
@@ -41,11 +41,11 @@ module.exports = {
 					allowNull: false,
 				},
 				lat: {
-					type: Sequelize.DECIMAL,
+					type: Sequelize.FLOAT,
 					allowNull: false,
 				},
 				lng: {
-					type: Sequelize.DECIMAL,
+					type: Sequelize.FLOAT,
 					allowNull: false,
 				},
 				name: {
@@ -57,7 +57,7 @@ module.exports = {
 					allowNull: false,
 				},
 				price: {
-					type: Sequelize.DECIMAL,
+					type: Sequelize.FLOAT,
 					allowNull: false,
 				},
 				createdAt: {
